refactor: Create SponsorCard class
