+++ lib/ui/screen/sessions/sessions.dart
@@ -1,9 +1,8 @@
 import 'package:conference_2023/l10n/localization.dart';
-import 'package:conference_2023/model/app_locale.dart';
-import 'package:conference_2023/model/favorites/favorite_session_ids.dart';
 import 'package:conference_2023/model/sessions/session.dart';
 import 'package:conference_2023/model/sessions/session_provider.dart';
 import 'package:conference_2023/ui/router/router_app.dart';
+import 'package:conference_2023/ui/widget/session_card.dart';
 import 'package:conference_2023/util/extension/build_context_ext.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -109,8 +108,9 @@ class _SessionSection extends ConsumerWidget {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              _SessionCard(
+              SessionCard(
                 session: session,
+                showFavoriteIcon: true,
               ),
               const Gap(4),
               Padding(
@@ -131,77 +131,3 @@ class _SessionSection extends ConsumerWidget {
   }
 }
 
-class _SessionCard extends ConsumerWidget {
-  const _SessionCard({
-    required this.session,
-  });
-
-  final Session session;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final locale = ref.watch(appLocaleProvider);
-    final favoriteSessionIds = ref.watch(favoriteSessionIdsNotifierProvider);
-
-    final speakerName = switch (session) {
-      SessionSponsor(speaker: final s) => Wrap(
-          spacing: 8,
-          crossAxisAlignment: WrapCrossAlignment.center,
-          children: [
-            Text(s.name),
-            const Icon(Icons.handshake_outlined),
-          ],
-        ),
-      SessionTalk(speaker: final s) => Text(s.name),
-      _ => null,
-    };
-
-    final leadingImage = switch (session) {
-      SessionSponsor(speaker: final s) ||
-      SessionTalk(speaker: final s) =>
-        CircleAvatar(
-          backgroundImage: NetworkImage(s.avatarUrl),
-        ),
-      _ => null,
-    };
-
-    final trailingFavorite = switch (session) {
-      SessionSponsor(id: final i) || SessionTalk(id: final i) => Icon(
-          favoriteSessionIds.contains(i)
-              ? Icons.favorite
-              : Icons.favorite_border,
-        ),
-      _ => null,
-    };
-
-    return Card(
-      elevation: 0,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-        side: BorderSide(
-          color: Theme.of(context).colorScheme.outline,
-        ),
-      ),
-      child: ListTile(
-        title: Text(session.title.get(locale)),
-        contentPadding: const EdgeInsets.symmetric(
-          horizontal: 16,
-        ),
-        subtitle: speakerName,
-        leading: leadingImage,
-        trailing: trailingFavorite,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-        ),
-        onTap: switch (session) {
-          SessionTalk() || SessionSponsor() => () {
-              SessionDetailRoute(
-                sessionId: session.id,
-              ).push(context);
-            },
-          _ => null,
-        },
-      ),
-    );
-  }
-}
